+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Adds support for XYZM coordinate types when unmarshalling GeoJSON.
+
 ## v0.54.0
 
 2025-06-16